Move db/ files into backend/
