+++ bigchaindb/backend/__init__.py
@@ -1,1 +1,4 @@
-from bigchaindb.backend.connection import connect  # noqa
+from bigchaindb.db.factory import get_backend_factory
+from bigchaindb.db.query import Query
+from bigchaindb.db.schema import Schema
+from bigchaindb.db.connection import Connection
