+++ bigchaindb/backend/schema.py
@@ -1,0 +1,13 @@
+class Schema:
+
+    def create_database(self):
+        raise NotImplementedError()
+
+    def create_tables(self):
+        raise NotImplementedError()
+
+    def create_indexes(self):
+        raise NotImplementedError()
+
+    def drop_database(self):
+        raise NotImplementedError()
