--- bigchaindb/db/__init__.py
@@ -1,4 +1,0 @@
-from bigchaindb.db.factory import get_backend_factory
-from bigchaindb.db.query import Query
-from bigchaindb.db.schema import Schema
-from bigchaindb.db.connection import Connection
