--- bigchaindb/db/connection.py
@@ -1,11 +1,0 @@
-class ConnectionError(Exception):
-    """Raised when there is a connection error when running a query."""
-
-
-class Connection:
-
-    def run(self):
-        raise NotImplementedError()
-
-    def connect(self):
-        raise NotImplementedError()
